+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.2.0 (in development)
 -----------------------
 - Drop support for Python 3.3
+- Fix a DeprecationWarning in Python 3.5+
 
 v0.1.0 (2017-10-02)
 -------------------